Initial ROS drake package and repository information
New ROS drake package based on just calling ament_vendor
to download, build and install the version 1.24 of Drake.

In this initial phase the GUROBI, MOSEK and SNOPT
components are disabled in the build using the CMake
parameters.

The ament_vendor uses a couple of patches to deal with
issues related to clang version and to disable the
assumption of platform having a libjchart2d-java package.
Patches uses the Debian DEP-3 format to include metadata
on the same file.

Dependencies are based on the official documentation and
comments are in package.xml.

Signed-off-by: Jose Luis Rivero <jrivero@osrfoundation.org>